Accuracy settings added Minor fix in PathBuilder

+++ TravellerDiary/TravellerDiary/Internal/Model/Path.h
@@ -9,7 +9,7 @@
 @property NSInteger userId;
 @property NSString *name;
 @property NSString *comment;
-@property NSInteger rating;
+@property float rating;
 @property (readonly) RLMLinkingObjects *coordinates;
 @property NSDate *createdAt;
 @property NSDate *updatedAt;
